requirements.txt
asgiref==3.6.0
crispy-bootstrap5==0.7
Django==4.1.6
django-crispy-forms==2.0
django-phonenumber-field==7.0.2
django-seed==0.3.1
djangorestframework==3.14.0
Faker==17.6.0
phonenumberslite==8.13.6
Pillow==9.4.0
psycopg2==2.9.5
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
toposort==1.10
tzdata==2022.7
